Move lib file to lib folder and vite lib build
