+++ platforms/web/src/.eslintignore
@@ -1,1 +1,3 @@
 src/images
+generated
+dist
