+++ platforms/web/src/lib/assert.ts
