+++ platforms/web/src/lib/useFormattingActions.ts
