+++ platforms/web/src/lib/useTestCases/useTestCases.ts
@@ -1,6 +1,6 @@
 import { RefObject, useCallback, useMemo, useRef, useState } from 'react';
 
-import { ComposerModel } from '../../../generated/wysiwyg';
+import { ComposerModel } from '../../generated/wysiwyg';
 import { Actions } from './types';
 import { getSelectionAccordingToActions, resetTestCase, traceAction } from './utils';
