+++ platforms/web/src/package.json
@@ -1,19 +1,29 @@
 {
-  "name": "matrix-wysiwsyg",
+  "name": "matrix-wysiwyg",
   "version": "0.0.0",
   "type": "module",
+  "description": "Wysiwyg composer for matrix.org using React",
   "author": "matrix.org",
   "repository": {
     "type": "git",
     "url": "https://github.com/matrix-org/matrix-wysiwyg"
   },
   "license": "Apache-2.0",
+  "main": "./dist/matrix-wysiwyg.umd.cjs",
+  "module": "./dist/matrix-wysiwyg.js",
+  "exports": {
+    ".": {
+      "import": "./dist/matrix-wysiwyg.js",
+      "require": "./dist/matrix-wysiwyg.umd.cjs"
+    }
+  },
+  "types": "./dist/index.d.ts",
   "scripts": {
     "dev": "vite",
-    "build": "tsc && vite build",
+    "build": "vite build",
     "preview": "vite preview",
     "typecheck": "tsc --noEmit",
-    "lint": "eslint ./src",
+    "lint": "eslint",
     "test": "vitest --run",
     "test:watch": "vitest",
     "test:ui": "vitest --ui",
@@ -44,6 +54,7 @@
     "react-testing-library": "^8.0.1",
     "typescript": "^4.7.4",
     "vite": "^3.1.0",
+    "vite-plugin-dts": "^1.5.0",
     "vitest": "^0.23.4"
   }
 }
