+++ platforms/web/src/src/App.tsx
@@ -1,6 +1,6 @@
 import { MouseEventHandler } from 'react';
 
-import { useWysiwyg } from './lib/useWysiwyg';
+import { useWysiwyg } from '../lib/useWysiwyg';
 import boldImage from './images/bold.svg';
 import undoImage from './images/undo.svg';
 import redoImage from './images/redo.svg';
