+++ platforms/web/src/vite.config.ts
@@ -1,19 +1,45 @@
 import { defineConfig } from 'vitest/config';
 import react from '@vitejs/plugin-react';
+import { resolve } from 'path';
+import dts from 'vite-plugin-dts';
 
 // https://vitejs.dev/config/
 export default defineConfig({
     // eslint-disable-next-line @typescript-eslint/ban-ts-comment
     // @ts-ignore
-    plugins: [react()],
+    plugins: [react(), dts({
+        include: ['lib/useWysiwyg.ts'],
+        insertTypesEntry: true,
+    })],
     test: {
         globals: true,
         environment: 'jsdom',
         setupFiles: 'test.setup.ts',
-        includeSource: ['src/**/*.{ts,tsx}'],
+        includeSource: ['lib/**/*.{ts,tsx}'],
         coverage: {
             all: true,
-            include: ['src/**/*.{ts,tsx}'],
+            include: ['lib/**/*.{ts,tsx}'],
+        },
+    },
+    build: {
+        lib: {
+            entry: resolve(__dirname, 'lib/useWysiwyg.ts'),
+            name: 'Matrix wysiwyg',
+            // the proper extensions will be added
+            fileName: 'matrix-wysiwyg',
+        },
+        rollupOptions: {
+            // make sure to externalize deps that shouldn't be bundled
+            // into your library
+            external: ['react', 'react-dom'],
+            output: {
+            // Provide global variables to use in the UMD build
+            // for externalized deps
+                globals: {
+                    'react': 'React',
+                    'react-dom': 'ReactDOM',
+                },
+            },
         },
     },
 });
